fix showDiff only working for ApexComponents
