+++ src/commands/showDiffCommand.ts
@@ -1,8 +1,7 @@
 import * as vscode from 'vscode';
 
 import {ICommand} from './command';
-import {getFileNameFromUri, getSalesforceTypeFromFileName} from '../utils/utils';
-import {SalesforceQueryBuilder} from '../builder/salesforceQueryBuilder';
+import {getFileNameFromUri, getSalesforceTypeFromFileName, buildSalesforceUriFromLocalUri} from '../utils/utils';
 
 
 /**
@@ -25,11 +24,24 @@ export class ShowDiffCommand implements ICommand {
    */
   public Execute(uri: vscode.Uri) {
     let filename: string = getFileNameFromUri(uri);
-    // let uri = new vscode.Uri();
-    // uri.
+/*
+    // Create a status bar item with Salesforce information about the last user that edited the page
+    let status = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right);
+
+    status.text = `$(git-commit) Last modified by ${results.records[0].LastModifiedBy.Name} on ${results.records[0].LastModifiedDate.substr(0, 10)}`;
+    status.tooltip = `Modifications done on the file in your Salesforce organization`;
+    status.show();
+
+    // Remove the status bar item when we are not in a diff (salesforce) mode
+    let event = vscode.workspace.onDidCloseTextDocument(() => {
+      status.dispose();
+      event.dispose();
+    });
+*/
+
     vscode.commands
       .executeCommand('vscode.diff',
       uri,
-      vscode.Uri.parse(), `${filename} (LOCAL) <~> ${filename} (REMOTE)`);
+      buildSalesforceUriFromLocalUri(uri), `${filename} (LOCAL) <~> ${filename} (REMOTE)`);
   }
 }
